+++ core/src/main/scala/com/raphtory/internals/graph/GraphAlteration.scala
@@ -5,7 +5,7 @@ import com.raphtory.api.input.Type
 import com.raphtory.internals.components.querymanager._
 
 sealed private[raphtory] trait GraphAlteration {
-  val sourceID: Int
+  val sourceID: Long
 }
 
 private[raphtory] object GraphAlteration {
@@ -18,7 +18,7 @@ private[raphtory] object GraphAlteration {
 
   /** basic update types */
   case class VertexAdd(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -27,12 +27,12 @@ private[raphtory] object GraphAlteration {
   )(implicit val provider: SchemaProvider[VertexAdd])
           extends GraphUpdate //add a vertex (or add/update a property to an existing vertex)
 
-  case class VertexDelete(sourceID: Int, updateTime: Long, index: Long, srcId: Long)(implicit
+  case class VertexDelete(sourceID: Long, updateTime: Long, index: Long, srcId: Long)(implicit
       val provider: SchemaProvider[VertexDelete]
   ) extends GraphUpdate
 
   case class EdgeAdd(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -42,7 +42,7 @@ private[raphtory] object GraphAlteration {
   )(implicit val provider: SchemaProvider[EdgeAdd])
           extends GraphUpdate
 
-  case class EdgeDelete(sourceID: Int, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
+  case class EdgeDelete(sourceID: Long, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
       val provider: SchemaProvider[EdgeDelete]
   ) extends GraphUpdate
 
@@ -53,7 +53,7 @@ private[raphtory] object GraphAlteration {
   }
 
   case class SyncNewEdgeAdd(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -66,7 +66,7 @@ private[raphtory] object GraphAlteration {
   ) extends GraphUpdateEffect(dstId)
 
   case class BatchAddRemoteEdge(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -79,7 +79,7 @@ private[raphtory] object GraphAlteration {
           with GraphUpdate
 
   case class SyncExistingEdgeAdd(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -89,12 +89,12 @@ private[raphtory] object GraphAlteration {
       val provider: SchemaProvider[SyncExistingEdgeAdd]
   ) extends GraphUpdateEffect(dstId)
 
-  case class SyncExistingEdgeRemoval(sourceID: Int,updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
+  case class SyncExistingEdgeRemoval(sourceID: Long, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
       val provider: SchemaProvider[SyncExistingEdgeRemoval]
   ) extends GraphUpdateEffect(dstId)
 
   case class SyncNewEdgeRemoval(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -105,17 +105,17 @@ private[raphtory] object GraphAlteration {
   ) extends GraphUpdateEffect(dstId)
 
   /** Edge removals generated via vertex removals */
-  case class OutboundEdgeRemovalViaVertex(sourceID: Int, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
+  case class OutboundEdgeRemovalViaVertex(sourceID: Long, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
       val provider: SchemaProvider[OutboundEdgeRemovalViaVertex]
   ) extends GraphUpdateEffect(dstId)
 
-  case class InboundEdgeRemovalViaVertex(sourceID: Int, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
+  case class InboundEdgeRemovalViaVertex(sourceID: Long, updateTime: Long, index: Long, srcId: Long, dstId: Long)(implicit
       val provider: SchemaProvider[InboundEdgeRemovalViaVertex]
   ) extends GraphUpdateEffect(srcId)
 
   /** Responses from a partition receiving any of the above */
   case class SyncExistingRemovals(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -127,7 +127,7 @@ private[raphtory] object GraphAlteration {
   ) extends GraphUpdateEffect(srcId)
 
   case class EdgeSyncAck(
-      sourceID: Int,
+      sourceID: Long,
       updateTime: Long,
       index: Long,
       srcId: Long,
@@ -137,7 +137,7 @@ private[raphtory] object GraphAlteration {
       val provider: SchemaProvider[EdgeSyncAck]
   ) extends GraphUpdateEffect(srcId)
 
-  case class VertexRemoveSyncAck(sourceID: Int, updateTime: Long, index: Long, override val updateId: Long)(implicit
+  case class VertexRemoveSyncAck(sourceID: Long, updateTime: Long, index: Long, override val updateId: Long)(implicit
       val provider: SchemaProvider[VertexRemoveSyncAck]
   ) extends GraphUpdateEffect(updateId)
 }
